test: move to internal/test
