+++ Makefile
@@ -34,9 +34,9 @@ binaries:
 
 ci:
 	# task used by CI
-	go install ./cmd/trace-agent
 	go get -u golang.org/x/lint/golint
-	golint -set_exit_status=1 ./test ./cmd/trace-agent ./internal/filters ./internal/api ./internal/testutil ./internal/info ./internal/quantile ./internal/obfuscate ./internal/sampler ./internal/statsd ./internal/watchdog ./internal/writer ./internal/flags ./internal/osutil
+	golint -set_exit_status=1 ./cmd/trace-agent ./internal/filters ./internal/api ./internal/test ./internal/info ./internal/quantile ./internal/obfuscate ./internal/sampler ./internal/statsd ./internal/watchdog ./internal/writer ./internal/flags ./internal/osutil
+	go install ./cmd/trace-agent
 	go test -v -race ./...
 
 windows:
