+++ cmd/trace-agent/agent_test.go
@@ -21,7 +21,7 @@ import (
 	"github.com/DataDog/datadog-trace-agent/internal/info"
 	"github.com/DataDog/datadog-trace-agent/internal/obfuscate"
 	"github.com/DataDog/datadog-trace-agent/internal/sampler"
-	"github.com/DataDog/datadog-trace-agent/internal/testutil"
+	"github.com/DataDog/datadog-trace-agent/internal/test/testutil"
 )
 
 type mockSamplerEngine struct {
