+++ internal/test/buffer_test.go
