+++ internal/writer/fixtures_test.go
@@ -4,7 +4,7 @@ import (
 	"math/rand"
 	"sync"
 
-	"github.com/DataDog/datadog-trace-agent/internal/testutil"
+	"github.com/DataDog/datadog-trace-agent/internal/test/testutil"
 )
 
 // payloadConstructedHandlerArgs encodes the arguments passed to a PayloadConstructedHandler call.
